Rename example app to UplinkDemo from console
