+++ app/build.gradle
@@ -6,7 +6,7 @@ android {
     compileSdk 31
 
     defaultConfig {
-        applicationId "io.bytebeam.console"
+        applicationId "io.bytebeam.UplinkDemo"
         minSdk 21
         targetSdk 31
         versionCode 1
